% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{gibbsC}
\alias{gibbsC}
\title{A Gibbs sampler using Rcpp}
\usage{
gibbsC(N, n, a, b)
}
\arguments{
\item{N}{the number of samples}

\item{n}{one parameter of  bivariate density}

\item{a}{one parameter of  bivariate density}

\item{b}{one parameter of  bivariate density}
}
\value{
a random sample of size \code{n}
}
\description{
A Gibbs sampler using Rcpp
}
\examples{
\dontrun{
rnC <- gibbsC(100,10,2,3)
}
}
