% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLS.R
\name{sml}
\alias{sml}
\title{simulation of tls when dim of regressor is two}
\usage{
sml(t, n, beta1, beta2, delta, gam)
}
\arguments{
\item{t}{time of repitition}

\item{n}{numbers of random}

\item{beta1}{slope coefficient for the spatial lag term if threshold vaviable is less than threshold parameter}

\item{beta2}{slope coefficient for the spatial lag term if threshold vaviable is larger than threshold parameter}

\item{delta}{common slope parameters}

\item{gam}{threshold parameter}
}
\value{
A table of 5%, 50% and 95% of the quantiles of distributions of the estimators of parameters
}
\description{
simulation of tls when dim of regressor is two
}
\examples{
\dontrun{
 t<-1000;n<-100;
 gamma<-2;beta1<-0.2;beta2<-0.3;delta<-c(1,1);
 sml(t,n,beta1,beta2,delta,gamma)
}
}
