\name{StatComp21011-package}
\alias{StatComp21011-package}
\alias{StatComp21011}
\docType{package}
\title{
  This package estimates the threshold paramater for spatial threshold autoregressive model
}
\description{
  two stage estimation method for the threshold paramater for spatial threshold autoregressive model
}
\details{
  This section should provide a more detailed overview of how to use the
  package, including the most important functions.
}
\author{
Xinrong Xiang

Maintainer: Xinrong Xiang <xxrlymxy@126.com>
}
\references{
  This optional section can contain literature or other references for
  background information.
}
\keyword{ package }
\seealso{
  Optional links to other man pages
}
\examples{
  \dontrun{
     ## Optional simple examples of the most important functions
     ## These can be in \dontrun{} and \donttest{} blocks.   
  }
}
